[No Need Upstreaming] fix error message (#231)
(cherry picked from commit 2e495a67755d27569940e347aa10727e9acca871)
(cherry picked from commit 259e6de6ea88bd661bf4b4cd776b9e6e091e0fc6)

+++ c10/xpu/XPUFunctions.cpp
@@ -113,9 +113,9 @@ inline void check_device(DeviceIndex device) {
   TORCH_CHECK(
       device >= 0 && device < total,
       "device is out of range, device is ",
-      device,
+      static_cast<int16_t>(device),
       ", total number of device is ",
-      total,
+      static_cast<int16_t>(total),
       ".");
 }
